added pattern test cases for detectors starting with b
